Bug 1674: Wrong error message when attempting to rename an S3 folder
https://winscp.net/tracker/1674
(cherry picked from commit 44a60305862ff17f4c5ae78008228cc3a0921db2)

Source commit: 5b2a87cb28da63f5b9eae5c00fc69f6a51177608

+++ source/core/S3FileSystem.cpp
@@ -965,6 +965,10 @@ void __fastcall TS3FileSystem::DeleteFile(const UnicodeString AFileName,
 void __fastcall TS3FileSystem::RenameFile(const UnicodeString FileName, const TRemoteFile * File,
   const UnicodeString NewName)
 {
+  if (DebugAlwaysTrue(File != NULL) && File->IsDirectory)
+  {
+    throw Exception(LoadStr(NOTSUPPORTED));
+  }
   CopyFile(FileName, File, NewName);
   TRmSessionAction DummyAction(FTerminal->ActionLog, FileName);
   DeleteFile(FileName, File, dfForceDelete, DummyAction);
